Ada morse code equivalent
Since completing the C version, I was able to debug the issues I was having the original Ada version.  Binary sizes would be similar if not for the Ada compiler pulling in the entirety of Alphabet.ads data (where the C compiler know to pull just the characters used in the message). I figure this the difference between the "#define" macro and constant delcaration.